Rename file for consistency
